+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_data.py
@@ -1592,6 +1592,29 @@ class WinsorizeCareHomeFilledPostsPerBedRatioOutliersData:
 
 
 @dataclass
+class ConvertPirPeopleToFilledPostsData:
+    valid_rows = [
+        (CareHome.care_home,     10.0,   10.0),  # invalid (care home)
+        (CareHome.not_care_home, None,   10.0),  # invalid (people is null)
+        (CareHome.not_care_home, 0.0,    10.0),  # invalid (people is zero)
+        (CareHome.not_care_home, 10.0,   None),  # invalid (posts in null)
+        (CareHome.not_care_home, 10.0,    0.0),  # invalid (posts is zero)
+        (CareHome.not_care_home, 100.0,  10.0),  # invalid (ratio too low, outside abs_diff cutoff)
+        (CareHome.not_care_home, 10.0,  100.0),  # invalid (ratio too high, outside abs_diff cutoff)
+        (CareHome.not_care_home, 20.0,   35.0),  # valid (outside abs_diff cutoff but ratio ok)
+        (CareHome.not_care_home, 10.0,    1.0),  # valid (ratio too low but inside abs_diff cutoff)
+        (CareHome.not_care_home, 1.0,    10.0),  # valid (ratio too high but inside abs_diff cutoff)
+        (CareHome.not_care_home, 10.0,   10.0),  # valid (ratio and abs_diff within limits)
+    ] # fmt: skip
+    expected_valid_rows = [
+        (CareHome.not_care_home, 20.0,   35.0),
+        (CareHome.not_care_home, 10.0,    1.0),
+        (CareHome.not_care_home, 1.0,    10.0),
+        (CareHome.not_care_home, 10.0,   10.0),
+    ] # fmt: skip
+
+
+@dataclass
 class NullCtPostsToBedsOutliers:
     null_ct_posts_to_beds_outliers_rows = [
         ("1-001", 1, 1.00, 1, CTFilteringRule.populated),